+++ see/tools.py
@@ -4,6 +4,7 @@ Filtering and other tasks.
 """
 import fnmatch
 import re
+import unicodedata
 
 
 def compact(objects):
@@ -11,6 +12,29 @@ def compact(objects):
     Filter out any falsey objects in a sequence.
     """
     return tuple(filter(bool, objects or []))
+
+
+def char_width(char):
+    """
+    Get the display length of a unicode character.
+    """
+    if ord(char) < 128:
+        return 1
+    elif unicodedata.east_asian_width(char) in ('F', 'W'):
+        return 2
+    elif unicodedata.category(char) in ('Mn',):
+        return 0
+    else:
+        return 1
+
+
+def display_len(text):
+    """
+    Get the display length of a string. This can differ from the character
+    length if the string contains wide characters.
+    """
+    text = unicodedata.normalize('NFD', text)
+    return sum(char_width(char) for char in text)
 
 
 def filter_regex(names, pattern):
@@ -27,7 +51,7 @@ def filter_regex(names, pattern):
 
 def filter_wildcard(names, pattern):
     """
-    Return a tuple of strings that match a shell-style pattern.
+    Return a tuple of strings that match a shell-style wildcard pattern.
     """
     def match(name, fn=fnmatch.fnmatch, pattern=pattern):
         return fn(name, pattern)